pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dns-healthcheck"
version = "0.1.0"
description = "Comprehensive DNS auditor with DNSSEC, email security, web posture, and multi-resolver propagation checks"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "dns-healthcheck contributors" }]
keywords = ["dns", "dnssec", "audit", "security", "spf", "dmarc", "caa", "sarif", "healthcheck"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: Security",
    "Topic :: System :: Networking :: Monitoring",
]
dependencies = [
    "dnspython>=2.6.0",
    "cryptography>=42.0",
    "typer>=0.12",
    "rich>=13.7",
    "httpx>=0.27",
    "idna>=3.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-requests",
]

[project.scripts]
dnshc = "dns_healthcheck.cli:app"
dns-healthcheck = "dns_healthcheck.cli:app"

[project.urls]
Homepage = "https://github.com/Zv1r/dns-healthcheck"
Source = "https://github.com/Zv1r/dns-healthcheck"
Issues = "https://github.com/Zv1r/dns-healthcheck/issues"

[tool.hatch.build.targets.wheel]
packages = ["dns_healthcheck"]

[tool.hatch.build.targets.sdist]
include = ["/dns_healthcheck", "/tests", "/README.md", "/LICENSE"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "C4", "SIM", "PL", "RUF"]
ignore = [
    "PLR0913",   # too-many-arguments — intentional for check ctor signatures
    "PLR2004",   # magic-value-comparison — RFC numbers are clearer inline
    "PLR0912",   # too-many-branches — checks legitimately fan out
    "PLR0915",   # too-many-statements
    "PLR0911",   # too-many-return-statements
    "PLC0415",   # import-outside-top-level — used for lazy DNS submodules
    "B008",      # function-call-in-default-argument — Typer pattern
    "SIM117",    # nested-with — clearer for sockets
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR2004", "S101"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["dns_healthcheck"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
